style: 전역 reset, blobal style 연결
