+++ week4/package.json
@@ -12,7 +12,8 @@
   "dependencies": {
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.18.0"
+    "react-router-dom": "^6.18.0",
+    "styled-reset": "^4.5.1"
   },
   "devDependencies": {
     "@types/react": "^18.2.15",
