--- week4/src/sytles/globalStyle.js
@@ -1,20 +1,0 @@
-const colors = {
-    white: "#ffffff",
-    gray1: "#F2F3F7",
-    gray2: "#4E515C",
-    yellow: "#faeD62",
-    black: "#101011"
-    
-
-
-};
-
-const font = {
-    spoka: font-family: 'Spoqa Han Sans Neo', 'sans-serif';
-};
-
-const theme = {
-    colors, font
-};
-
-export default theme;
